Shifting around exchange token logic
