+++ libraries/microsoft-agents-hosting-core/microsoft_agents/hosting/core/app/oauth/authorization.py
@@ -328,78 +328,6 @@ class Authorization:
         token = sign_in_state.tokens[auth_handler_id]
         return TokenResponse(token=token)
 
-    async def exchange_token(
-        self,
-        context: TurnContext,
-        scopes: list[str],
-        auth_handler_id: str,
-    ) -> TokenResponse:
-        """
-        Exchanges a token for another token with different scopes.
-
-        :param context: The context object for the current turn.
-        :type context: TurnContext
-        :param scopes: The scopes to request for the new token.
-        :type scopes: list[str]
-        :param auth_handler_id: Optional ID of the auth handler to use, defaults to first
-        :type auth_handler_id: str
-        :return: The token response from the OAuth provider from the exchange.
-            If the cached token is not exchangeable, returns the cached token.
-        :rtype: TokenResponse
-        """
-
-        token_response = await self.get_token(context, auth_handler_id)
-
-        if token_response and self._is_exchangeable(token_response.token):
-            logger.debug("Token is exchangeable, performing OBO flow")
-            return await self._handle_obo(token_response.token, scopes, auth_handler_id)
-
-        return token_response
-
-    def _is_exchangeable(self, token: str) -> bool:
-        """
-        Checks if a token is exchangeable (has api:// audience).
-
-        :param token: The token to check.
-        :type token: str
-        :return: True if the token is exchangeable, False otherwise.
-        """
-        try:
-            # Decode without verification to check the audience
-            payload = jwt.decode(token, options={"verify_signature": False})
-            aud = payload.get("aud")
-            return isinstance(aud, str) and aud.startswith("api://")
-        except Exception:
-            logger.error("Failed to decode token to check audience")
-            return False
-
-    async def _handle_obo(
-        self, token: str, scopes: list[str], handler_id: str = None
-    ) -> TokenResponse:
-        """
-        Handles On-Behalf-Of token exchange.
-
-        :param token: The original token.
-        :type token: str
-        :param scopes: The scopes to request.
-        :type scopes: list[str]
-        :param handler_id: The ID of the auth handler to use, defaults to first
-        :type handler_id: str, optional
-        :return: The new token response.
-        :rtype: TokenResponse
-        """
-        auth_handler = self.resolve_handler(handler_id)
-        token_provider = self._connection_manager.get_connection(
-            auth_handler.obo_connection_name
-        )
-
-        logger.info("Attempting to exchange token on behalf of user")
-        new_token = await token_provider.aquire_token_on_behalf_of(
-            scopes=scopes,
-            user_assertion=token,
-        )
-        return TokenResponse(token=new_token)
-
     def on_sign_in_success(
         self,
         handler: Callable[[TurnContext, TurnState, Optional[str]], Awaitable[None]],
