--- tests/_common/testing_objects/http/testing_client_session.py
@@ -1,2 +1,0 @@
-class TestingClientSessionBase:
-    pass
