+++ tests/hosting_core/app/test_agent_application.py
@@ -1,0 +1,36 @@
+from microsoft_agents.authentication.msal.msal_connection_manager import MsalConnectionManager
+from microsoft_agents.hosting.core.turn_context import TurnContext
+import pytest
+
+from microsoft_agents.authentication.msal import MsalAuthentication
+from microsoft_agents.hosting.core import (
+    MemoryStorage,
+    AgentApplication,
+    ApplicationOptions,
+    Connections
+)
+
+def mock_send_activity(mocker):
+    mocker.patch.object(TurnContext, 'send_activity', new=)
+
+class TestUtils:
+
+    @pytest.fixture
+    def options(self):
+        return ApplicationOptions()
+    
+    @pytest.fixture
+    def storage(self):
+        return MemoryStorage()
+    
+    @pytest.fixture
+    def connection_manager(self):
+        return MsalConnectionManager()
+    
+    @pytest.fixture
+    def 
+        
+
+class TestAgentApplication:
+
+    pass
